Factored out test code
